requirements.txt
Flask==3.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
jdatetime==4.1.1
openpyxl==3.1.2
gunicorn==21.2.0
